% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiColors.r
\name{guiBlue}
\alias{guiBlue}
\title{Blue colors}
\usage{
guiBlue(dark = FALSE)
}
\arguments{
\item{dark}{Switch between light and dark color. Logical.}
}
\value{
Character string representing a color.
}
\description{
Blue colors to be used with GUI.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
