% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDynamic.r
\name{showDynamic}
\alias{showDynamic}
\title{Default dynamics plots}
\usage{
showDynamic(sim, labelScenario, labelTime)
}
\arguments{
\item{sim}{Numeric array with three dimensions
(time, item, scenario). The names of the scenario dimension are
integers starting from 1.}

\item{labelScenario}{Word 'scenario' translated to target language
for use in the legend. Character string.}

\item{labelTime}{Word 'time' translated to target language
for use as x-asis label. Character string.}
}
\value{
A data frame with two columns 'label' and 'content'. The
  latter column holds HTML or SVG code.
}
\description{
Creates default plots from dynamic model output. One plor is created
for each output item (i.e. state variable or process rate) showing
the respective time series for all scenarios.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
