% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showEffect.r
\name{showEffect}
\alias{showEffect}
\title{Default effect plots}
\usage{
showEffect(sim)
}
\arguments{
\item{sim}{Numeric matrix (rows: output items, colums: values of
the input item whose effect was studied).}
}
\value{
A data frame with two columns 'label' and 'content'. The
  latter column holds HTML or SVG code.
}
\description{
Creates default plots from steady-state effect analysis. One plot is
created for each output item (i.e. state variable or process rate).
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
