% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSteady.r
\name{showSteady}
\alias{showSteady}
\title{Default steady-state visualization}
\usage{
showSteady(sim)
}
\arguments{
\item{sim}{Numeric matrix (rows: output items, colums: scenarios).}
}
\value{
A data frame with two columns 'label' and 'content'. The
  latter column holds HTML or SVG code.
  This is currently a single-row data frame.
}
\description{
Creates a HTML table illustrating the output of a steady state
simulation. There is one row for each output item (i.e.
state variable or process rate) and one column for each scenario.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
