% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{preGUI}
\alias{preGUI}
\title{Prepare files needed by the GUI}
\usage{
preGUI(appName = "rodeoGUI", vars = "vars.txt", pars = "pars.txt",
  funs = "funs.txt", pros = "pros.txt", stoi = "stoi.txt",
  sourcesR = "functions.r", sourcesF = "functions.f95",
  dirScenarios = "./scenarios", dirIntro = "./intro", colsep = "\\t",
  useTemp = TRUE)
}
\arguments{
\item{appName}{Application name displayed in top left corner of GUI.}

\item{vars}{Input table.}

\item{pars}{Input table.}

\item{funs}{Input table.}

\item{pros}{Input table.}

\item{stoi}{Input table.}

\item{sourcesR}{R source files.}

\item{sourcesF}{Fortran source files.}

\item{dirScenarios}{Directory containing scenario definitions. See notes.}

\item{dirIntro}{Directory containing material for the
model's introduction pages. See notes.}

\item{colsep}{Column separator used in delimited text files.}

\item{useTemp}{If \code{TRUE} (default), the produced files are created in
the session's temporary folder. If \code{FALSE}, the files are creared in
the current working directory (which is rarely useful if the GUI is to be
run on the local machine).}
}
\value{
The function does not return anything but it creates two files in the
  current working directory. The first file with the fixed name
  'rodeoGuiData.rda' holds various data needed by the GUI. The second file is
  a system-specific shared library. The file name starts with 'rodeo'
  followed by a random pattern and the extension according to the
  \code{dynlib.ext} component of \code{\link[base]{.Platform}}.
}
\description{
Prepare files needed to run a rodeo-based model in the shiny GUI.
}
\note{
Detailed help text still missing.
}
\examples{
# see 'runGUI'
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
