% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{Locally run rodeo-based model in GUI}
\usage{
runGUI(useTemp = TRUE)
}
\arguments{
\item{useTemp}{If \code{TRUE} (default), the function expects to find the
two files created by \code{preGUI} in the sessions's temporary folder. If
\code{FALSE}, the function attempts to find the files in the
current working directory.}
}
\value{
\code{NULL}
}
\description{
Locally run a rodeo-based model in a specialized shiny GUI.
}
\examples{
\dontrun{
library(rodeoGUI)
preGUI(
  dirRodeo=system.file("examples/tank2/rodeo", package="rodeoGUI"),
  dirScenarios=system.file("examples/tank2/scenarios", package="rodeoGUI"),
  dirIntro= =system.file("examples/tank2/intro", package="rodeoGUI"),
  colsep="\\t"
)
runGUI()
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
