% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDynamic.r
\name{showDynamic}
\alias{showDynamic}
\title{Default dynamics plots}
\usage{
showDynamic(sim, prm, lang)
}
\arguments{
\item{sim}{Numeric array with three dimensions
(time, item, scenario).}

\item{prm}{Numeric matrix (rows: parameters, columns: scenarios).
This is available just for the case that the parameter values are
needed for visualization.}

\item{lang}{Identifier of selected language (character string).}
}
\value{
A data frame with two columns 'label' and 'content'. The
  latter column holds HTML or SVG code.
}
\description{
Creates default plots from dynamic model output. One plor is created
for each output item (i.e. state variable or process rate) showing
the respective time series for all scenarios.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
